Move packages around to split between v1 work and v2 work
