+++ dev/.rat-excludes
@@ -103,4 +103,4 @@ org.apache.spark.scheduler.ExternalClusterManager
 org.apache.spark.deploy.yarn.security.ServiceCredentialProvider
 spark-warehouse
 structured-streaming/*
-org.apache.spark.deploy.rest.kubernetes.DriverServiceManager
+org.apache.spark.deploy.kubernetes.submit.v1.DriverServiceManager
