+++ resource-managers/kubernetes/core/src/main/resources/META-INF/services/org.apache.spark.deploy.kubernetes.submit.v1.DriverServiceManager
@@ -1,0 +1,2 @@
+org.apache.spark.deploy.kubernetes.submit.v1.ExternalSuppliedUrisDriverServiceManager
+org.apache.spark.deploy.kubernetes.submit.v1.NodePortUrisDriverServiceManager
