--- resource-managers/kubernetes/core/src/main/resources/META-INF/services/org.apache.spark.deploy.rest.kubernetes.DriverServiceManager
@@ -1,2 +1,0 @@
-org.apache.spark.deploy.rest.kubernetes.ExternalSuppliedUrisDriverServiceManager
-org.apache.spark.deploy.rest.kubernetes.NodePortUrisDriverServiceManager
