+++ resource-managers/kubernetes/core/src/main/scala/org/apache/spark/deploy/rest/kubernetes/v1/PemsToKeyStoreConverter.scala
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.deploy.rest.kubernetes
+package org.apache.spark.deploy.rest.kubernetes.v1
 
 import java.io.{File, FileInputStream, FileOutputStream, InputStreamReader}
 import java.nio.file.Paths
