+++ resource-managers/kubernetes/core/src/main/scala/org/apache/spark/deploy/rest/kubernetes/v2/KubernetesSparkDependencyServer.scala
@@ -59,3 +59,9 @@ private[spark] class KubernetesSparkDependencyServer(
     jettyServer = None
   }
 }
+
+object KubernetesSparkDependencyServer {
+  def main(args: Array[String]): Unit = {
+    
+  }
+}
